rag/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.15.0
faiss-cpu==1.12.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-huggingface==0.3.1
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.31
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
olefile==0.47
omegaconf==2.3.0
openai==1.109.1
orjson==3.11.3
packaging==25.0
pillow==11.3.0
propcache==0.3.2
psutil==7.1.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pypdf==6.1.1
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.5.0
webencodings==0.5.1
wrapt==1.17.3
yarl==1.20.1
zstandard==0.25.0
